requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
typing-extensions==4.7.1